nodepool-config-builder/requirements.txt
PyYAML==3.12
paramiko==2.4.1
